Move raw param reading into the negotiator
