+++ src/Capability/CapabilityNegotiator.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Firehed\PhpLsp\Capability;
 
 use Firehed\PhpLsp\Protocol\InitializeResult;
+use Firehed\PhpLsp\Protocol\MarkupKind;
 use Firehed\PhpLsp\Protocol\Message;
 use Firehed\PhpLsp\ServerInfo;
 
@@ -32,12 +33,55 @@ final class CapabilityNegotiator
 
     public function negotiate(Message $message): InitializeResult
     {
-        $this->sessionCapabilities = SessionCapabilities::fromMessage($message);
+        $this->sessionCapabilities = $this->resolveSessionCapabilities($message);
 
         return new InitializeResult(
             capabilities: $this->advertisedCapabilities(),
             serverInfo: $this->serverInfo,
         );
+    }
+
+    private function resolveSessionCapabilities(Message $message): SessionCapabilities
+    {
+        $capabilities = self::readMap($message->params ?? [], 'capabilities');
+        $textDocument = self::readMap($capabilities, 'textDocument');
+        $completionItem = self::readMap(self::readMap($textDocument, 'completion'), 'completionItem');
+
+        return new SessionCapabilities(
+            hoverMarkupKind: self::negotiateHoverMarkupKind(self::readMap($textDocument, 'hover')),
+            snippetSupport: ($completionItem['snippetSupport'] ?? false) === true,
+        );
+    }
+
+    /**
+     * Per [LSP] `HoverClientCapabilities`, `contentFormat` lists the kinds the
+     * client supports in the order it prefers them; the first the server also
+     * supports wins.
+     *
+     * @param array<array-key, mixed> $hover
+     */
+    private static function negotiateHoverMarkupKind(array $hover): MarkupKind
+    {
+        foreach (self::readMap($hover, 'contentFormat') as $format) {
+            $kind = is_string($format) ? MarkupKind::tryFrom($format) : null;
+            if ($kind !== null) {
+                return $kind;
+            }
+        }
+
+        return MarkupKind::PlainText;
+    }
+
+    /**
+     * @param array<array-key, mixed> $source
+     *
+     * @return array<array-key, mixed>
+     */
+    private static function readMap(array $source, string $key): array
+    {
+        $value = $source[$key] ?? null;
+
+        return is_array($value) ? $value : [];
     }
 
     /**
